requirements.txt
gradio==4.29.0
openai>=1.0.0
anthropic>=0.25.0
mistralai>=0.3.0
requests>=2.31.0
python-dotenv>=1.0.1
elevenlabs>=1.0.0,<2.0.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.9
ffmpeg-python>=0.2.0
numpy>=1.24.0
whisper-timestamped>=1.14.4
torch>=2.0.0
transformers>=4.36.0
pillow>=10.0.0
python-telegram-bot>=20.0
google-api-python-client>=2.0
google-auth>=2.0
google-auth-oauthlib>=0.4
google-auth-httplib2>=0.1
mkdocs>=1.5.0
mkdocs-material>=9.0